+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_list_app_templates_request.py
@@ -13,6 +13,8 @@ class ListAppTemplatesRequest(DaraModel):
         keyword: str = None,
         max_results: int = None,
         next_token: str = None,
+        order_column: str = None,
+        order_type: str = None,
         page_num: int = None,
         page_size: int = None,
         product_version: str = None,
@@ -32,11 +34,15 @@ class ListAppTemplatesRequest(DaraModel):
         self.max_results = max_results
         # The token for the next query. This parameter is empty if no more results exist.
         self.next_token = next_token
+        # The sort field. Only copyCount and gmtModified are supported.
+        self.order_column = order_column
+        # The sort direction. Valid values: ASC and DESC. Default value: DESC.
+        self.order_type = order_type
         # The page number.
         self.page_num = page_num
-        # The page size.
+        # The number of entries per page.
         self.page_size = page_size
-        # The Edition.
+        # The edition.
         self.product_version = product_version
         # The template status.
         self.status = status
@@ -66,6 +72,12 @@ class ListAppTemplatesRequest(DaraModel):
 
         if self.next_token is not None:
             result['NextToken'] = self.next_token
+
+        if self.order_column is not None:
+            result['OrderColumn'] = self.order_column
+
+        if self.order_type is not None:
+            result['OrderType'] = self.order_type
 
         if self.page_num is not None:
             result['PageNum'] = self.page_num
@@ -100,6 +112,12 @@ class ListAppTemplatesRequest(DaraModel):
 
         if m.get('NextToken') is not None:
             self.next_token = m.get('NextToken')
+
+        if m.get('OrderColumn') is not None:
+            self.order_column = m.get('OrderColumn')
+
+        if m.get('OrderType') is not None:
+            self.order_type = m.get('OrderType')
 
         if m.get('PageNum') is not None:
             self.page_num = m.get('PageNum')